build: Remove "|& tee" everywhere
The problem with using a pipe is that it hides the error from the
original command.  For example if you have a make rule such as:

  foo.o:
    gcc -c foo.c

if the compilation of gcc fails that will cause the Makefile to stop
(which is what you want).  However if you change this to:

  foo.o:
    gcc -c foo.c |& tee -a log

then the pipe hides the error from gcc (instead returning the error
from the right hand side, tee command in this case, which is unlikely
to ever fail).  This means the Makefile continues running after the
error.

If you want to log the output of make, do it when you invoke make, eg:

  make |& tee log

Signed-off-by: Richard W.M. Jones <rjones@redhat.com>