build: Don't fail in iconv subdirectory
Previously we ran make -j10 |& tee ... which hid the failure when
glibc builds.  In fact it builds libc.a OK (the target that we want),
but then fails a bit later when building in the iconv/ subdirectory
(which we don't care about for now).

It turns out we cannot do the simple fix of building target libc.a,
or at least I cannot work out what actual target to use for that.

Instead, ignore errors from the make command, but add an "ls libc.a"
(which will fail, and exit the whole build if libc.a was not built).

Signed-off-by: Richard W.M. Jones <rjones@redhat.com>