redis: Don't build with systemd
If you happen to have libsystemd installed (eg. systemd-devel on
Fedora) then redis detects this and adds systemd notify support[1].
However systemd static is not available, and so we cannot find the
symbol sd_notify (part of libsystemd, not a system call).  UKL
incorrectly renames this symbol as ukl_sd_notify and then the link
fails because the symbol cannot be found.

The easiest fix for this is to force redis to build without systemd
support, for now.

[1] https://www.freedesktop.org/software/systemd/man/sd_notify.html

Signed-off-by: Richard W.M. Jones <rjones@redhat.com>