Ignore failure of Nginx building during the build process
In order to implement TCP shortcuts in Nginx we have to extern the shortcut functions. This leads to an
undefined reference error to the shortcut functions during the build process which prevents the whole
build from building correctly. This patch allows the build process to ignore failure to make Nginx in
order for the build to build correctly.