Fix build for modern GCC
Modern gcc now attempts to generate position independent code unless it
is explicitly told not to. This conflicts with our usage of
-mcmodel=kernel.

For the glibc build, we need to pass -mcmodel=kernel and -mno-red-zone
as flags only to be used when building the static libraries.

For the hello world demo, we simply add -fno-PIC to the build CFLAGS.

Signed-off-by: Eric B Munson <munsoner@bu.edu>