Really clean, really
Leaving the build around in glibc-build and libgcc-build seems to cache
config (even when we run make clean there) so make sure we are clean by
deleting all the build artifacts. $PROGRAM clean should toast the local
copy of undefined_sys_hack.c if it was made by the build.

Signed-off-by: Eric B. Munson <munsoner@bu.edu>