Update UKL kernel to 6.3
We changed quite a bit about how UKL worked with the kernel as part of
the effort to make something we could submit to upstream. As part of the
rebase effort, we are moving to using those changes, but this requires
changes to the configuration script. We have new options that can be
changed, but the defaults should leave everything working as they did
prior.

This kernel does not yet have IST PF support or same stack (these will
come if there is desire for them).

Rather than build a disk image with supermin, we package a very simple
init ramdisk using Fedora 36 as a base (in a container) and pair that
with a init script that sets up the /proc /sys and /dev mounts then
execs the UKL process.

Signed-off-by: Eric B Munson <munsoner@bu.edu>