Make network configuration an init task
When using the kernel command line for network configuration we are
implicitly assuming that the plugged in interface will be the first
enumerated and therefore be called eth0. However, this is not the case
for most MOC machines. The init script now takes optional network
configuration parameters and can identify the specified interface name
from a given MAC address. We can also specify IP addresses at boot time
instead of requiring changes to the boot config.

Signed-off-by: Eric B Munson <munsoner@bu.edu>