Fix kernel config cleaning and memcached build
The kernel config should be cleaned when we clean the kernel or the
whole project because a rerun of the configure script may change
selected options that are then ignored if the config file is still
present.

The memcached build is guaranteed to fail at the link step, don't let
that stop the overall build process.

Signed-off-by: Eric B Munson <munsoner@bu.edu>