Fix locking and read/write cursor
The locking around transferring a connection was incorrect, letting the
receiving thread read from a protectd field without holding the lock.
Now, locking ensures that the read is protected as well.

Use the configured msg_size for read and write instead of what we got
from read to power write.

Signed-off-by: Eric B Munson <munsoner@bu.edu>