Big changes in the event handler base line
There were a number of outstanding bugs in the event handler base line
tcp echo client and server.

For the server:
Rework perf stat collection to be per CPU and have each thread generate
its own line in the output file.

For the client:
Stop measuring socket commands, we don't need those.
Ensure that our TSC logs are size with enough space.
Match the perf stats changes in the server

Signed-off-by: Eric B Munson <munsoner@bu.edu>