Fix perf output and collect earlier
Perf output is sometimes broken and it is due to allocating perf_sz
bytes for the string (instead of perf_sz +1) and then asking glibc to
interpret it as a string. Instead, use write to commit stats to disk.

Don't let the server sit idle while we write the transaction log. Do
perf collection first.

Signed-off-by: Eric B Munson <munsoner@bu.edu>